bugfix fir historical chat
